+++ spade-tests/src/ast_lowering.rs
@@ -1311,3 +1311,61 @@ snapshot_error! {
         impl T for _ {}
     "
 }
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_function_type,
+    "
+        fn test(x: int<{1+2}>) {}
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_function_return_type,
+    "
+        fn test() -> int<{1+2}> {}
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_struct_definition,
+    "
+        struct X {
+            x: int<{1+2}>
+        }
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_struct_port_definition,
+    "
+        struct port X {
+            x: int<{1+2}>
+        }
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_enum_definition,
+    "
+        enum X {
+            A{val: int<{1+2}>}
+        }
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_impl_trait,
+    "
+        struct X {}
+        trait T<#A> {}
+        impl T<{1+2}> for X {}
+    "
+}
+
+snapshot_error! {
+    const_generic_is_not_allowed_in_impl_target,
+    "
+        trait T {}
+        impl T for int<{1+2}> {}
+    "
+}